xcb: take abs of coordinates as the ints they fit
The span arithmetic passed the build's integer type to abs, which
takes an int: correct narrow, truncating wide. Device coordinates fit
an int on any build, so they are narrowed explicitly first.

Source code generated by Claude (Anthropic).
Commit message generated by Claude (Anthropic).

+++ src/lib/xpost_dev_xcb.c
@@ -538,12 +538,12 @@ int _fillrect(Xpost_Context *ctx,
     /* adjust ranges */
     if (width.int_.val < 0)
     {
-        width.int_.val = abs(width.int_.val);
+        width.int_.val = abs((int)(width.int_.val));
         x.int_.val -= width.int_.val;
     }
     if (height.int_.val < 0)
     {
-        height.int_.val = abs(height.int_.val);
+        height.int_.val = abs((int)(height.int_.val));
         y.int_.val -= height.int_.val;
     }
     if (x.int_.val < 0) x.int_.val = 0;
